Implement wp-fractional-scale-v1
This allows clients to provide buffers at native resolutions when
fractional scaling is used.

Virtual backend is adjusted to support scales as floats

+++ autotests/integration/internal_window.cpp
@@ -646,7 +646,7 @@ void InternalWindowTest::testScale()
     QMetaObject::invokeMethod(kwinApp()->platform(), "setVirtualOutputs", Qt::DirectConnection,
                               Q_ARG(int, 2),
                               Q_ARG(QVector<QRect>, QVector<QRect>({QRect(0, 0, 1280, 1024), QRect(1280 / 2, 0, 1280, 1024)})),
-                              Q_ARG(QVector<int>, QVector<int>({2, 2})));
+                              Q_ARG(QVector<qreal>, QVector<qreal>({2, 2})));
 
     QSignalSpy windowAddedSpy(workspace(), &Workspace::internalWindowAdded);
     HelperWindow win;
